mysite/requirement.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
Faker==6.6.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
